rag_backend/requirements.txt
django
djangorestframework
python-dotenv
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-groq
chromadb
sentence-transformers
pymupdf
pypdf
